Use native Tcell API and enable some enhancements.
This allows for richer color on terminals with 16 or more colors.
It also will work better on non-Unicode terminals.  Folks with
East Asian 8-bit terminals or GB18030 will see the same niceness
that UTF-8 users do.

This also allows this to work on more platforms, such as dragonfly
and solaris platforms.