Adding capitals, speed ctl, default ascii+kana mix
Adding capital letters to the alphaNumerics set.

Adding allTheCharacters - the combination of all the ascii and all the
kana's.  Making that the default.

Adding the '-k/--kana' option to select kana-only as before.

Adding the ability to set the character set to allTheCharacters
on-the-fly, via the 'b' key.

Adding the ability to control the FPS rate on-the-fly.  '+' increases
the rate by one, '-' decreases it by one, and '=' resets it to the
original speed.  Min and max are still 1 and 60, respectively.

Changed the function literal event handler to a labeled loop, per TODO
comment.